requirements.txt
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
requests==2.32.3
python-dateutil==2.9.0.post0
pytz==2024.1
moviepy==1.0.3
imageio==2.36.0
imageio-ffmpeg==0.5.1
Pillow==10.4.0
numpy==2.0.1
openai==1.44.0
